split mobile & desktop legacy footers into own files
