+++ packages/components/src/LegacyFooter/LegacyDesktopFooter.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react" />
+declare const LegacyDesktopFooter: () => JSX.Element;
+export default LegacyDesktopFooter;
