+++ packages/components/src/LegacyFooter/LegacyMobileFooter.d.ts
@@ -1,0 +1,3 @@
+/// <reference types="react" />
+declare const LegacyMobileFooter: () => JSX.Element;
+export default LegacyMobileFooter;
